Update EggIncGrind.java
this will work most likely. if yall are testing this and it doesn't work its probably because one of thet motors are a tetrix motor and it most likely needs to have an include file that is missing. PLEASE send me a picture if that is the case or if there are any more errors.