Update SkibidiDriveV4.java
This should be it; the final iteration. Integrated the arm file (EggIncGrind.java) and merged it with SkibidiDriveV2.java (this was the sole drive base with 0 arm code, so it was easy to implement there rather than in v3 or v4). As for the actual drive base since it worked fine, I didn't change it, we may need to remove diagonals if we can't get them to work on "Friyay" which is fine since we have strafing anyway. Apologies for this "Yapathon" I just thought it was fitting.